Fix for the first bug
Bug description:
In textual mode answers the user might include some accented characters
like "è" that causes the answer to be considered as wrong, In order to
fix this issue we mapped all these characters to their unaccented
equivalents. For example in this case "e".
A new class is created "IgnoredChar" that takes care of ignoring
accented characters both in user input and the correct answer if it has
any.